After re-formatting function arguments

+++ include/Utils.h
@@ -28,7 +28,7 @@ struct Conv_Layer_Info
 	unsigned int FiltW_Size;	// Filter width (FW).
 };
 
-const Conv_Layer_Info		Conv_Layer_Info_NULL = { 0, 0, 0, 0, 0, 0, 0 };
+const Conv_Layer_Info		Conv_Layer_Info_NULL = { (unsigned int)0, (unsigned int)0, (unsigned int)0, (unsigned int)0, (unsigned int)0, (unsigned int)0, (unsigned int)0 };
 
 struct MaxP_Layer_Info
 {
@@ -55,7 +55,7 @@ struct Data_Block_Info
 struct Data_Block
 {
 	DBID_t				ID				= DBID_t::Null();		// Data block ID.
-	DBID_t				Douplicate_of	= DBID_t::Null();		// Original block ID if this is a duplicate.
+	DBID_t				Douplicate_of	= DBID_t::Null			  ();		// Original block ID if this is a duplicate.
 	Data_Block_Types	DBT				= DBT_Null;				// Block type.
 	bool				ready			= false;				// Ready flag.
 	bool				alocated		= false;				// Allocation flag.
@@ -110,7 +110,7 @@ struct Scheduling_Node
 	bool					timmed			= false;				// Timing flag.
 	Conv_Layer_Info			CLI				= Conv_Layer_Info_NULL;	// The layer info
 	SNID_t					ID				= SNID_t::Null();		// Node ID.
-	SNID_t					Douplicate_of	= SNID_t::Null();		// Original node if duplicated.
+	SNID_t					Douplicate_of	= SNID_t::Null			  ();		// Original node if duplicated.
 	size_t					layer			= 0;					// Owning layer index.
 	Scheduling_info			Sch_Info		= Scheduling_info_NULL;	// Scheduling info (level and spacing).
 	Scheduling_types		type			= Sch_Null;				// Node type.
@@ -256,47 +256,109 @@ struct Mapping_Result
 Data_Block assign(Data_Block_Info inp);
 
 // Build Conv_Layer_Info from individual dimensions.
-Conv_Layer_Info assign(unsigned int B, unsigned int K, unsigned int C, unsigned int W, unsigned int H, unsigned int FH, unsigned int FW);
+Conv_Layer_Info assign							  (
+														unsigned int B,
+														unsigned int K,
+														unsigned int C,
+														unsigned int W,
+														unsigned int H,
+														unsigned int FH,
+														unsigned int FW);
 
 // Compare Conv_Layer_Info for B/H/W equality and K==C consistency.
-bool Eq_comapre_BHW_K1C2(Conv_Layer_Info i1, Conv_Layer_Info i2);
+bool Eq_comapre_BHW_K1C2							(
+														Conv_Layer_Info i1,
+														Conv_Layer_Info i2);
 
 // Print Conv_Layer_Info to stdout.
-void print(Conv_Layer_Info inp);
+void print										  (
+														Conv_Layer_Info inp);
 
 // Print a Dependency to stdout.
-void print(Dependency dep);
+void print										  (
+														Dependency dep);
 
 // Allocate a 7D array of DBID_t with sizes d0..d6.
-DBID_t******* new_DBID_7D	(unsigned int d0, unsigned int d1, unsigned int d2, unsigned int d3, unsigned int d4, unsigned int d5, unsigned int d6);
+DBID_t******* new_DBID_7D						   (
+														unsigned int d0,
+														unsigned int d1,
+														unsigned int d2,
+														unsigned int d3,
+														unsigned int d4,
+														unsigned int d5,
+														unsigned int d6);
 
 // Allocate a 7D array of SNID_t with sizes d0..d6.
-SNID_t******* new_SNID_7D	(unsigned int d0, unsigned int d1, unsigned int d2, unsigned int d3, unsigned int d4, unsigned int d5, unsigned int d6);
+SNID_t******* new_SNID_7D						   (
+														unsigned int d0,
+														unsigned int d1,
+														unsigned int d2,
+														unsigned int d3,
+														unsigned int d4,
+														unsigned int d5,
+														unsigned int d6);
 
 // Allocate a 4D array of DBID_t with sizes d0..d3.
-DBID_t****    new_DBID_4D	(unsigned int d0, unsigned int d1, unsigned int d2, unsigned int d3);
+DBID_t****    new_DBID_4D						   (
+														unsigned int d0,
+														unsigned int d1,
+														unsigned int d2,
+														unsigned int d3);
 
 // Allocate a 4D array of SNID_t with sizes d0..d3.
-SNID_t****    new_SNID_4D	(unsigned int d0, unsigned int d1, unsigned int d2, unsigned int d3);
+SNID_t****    new_SNID_4D						   (
+														unsigned int d0,
+														unsigned int d1,
+														unsigned int d2,
+														unsigned int d3);
 
 // Construct a Dependency with given satisfied flag and dependent ID.
-Dependency Gen_Dependency(bool satisfied, SNID_t Dependent_ID);
+Dependency Gen_Dependency						   (
+														bool satisfied,
+														SNID_t Dependent_ID);
 
 // Map baseline count to vault count (heuristic).
-size_t calculate_vault_map(size_t bl);
+size_t calculate_vault_map						  (
+														size_t bl);
 
 // Return a string for a boolean result ("Suucessfull"/"Unsuucessfull").
-std::string Res_Dec(const bool b);
+std::string Res_Dec								 (
+														const bool b);
 
 // embedding the level value and baseline value			to the template
-std::string embedd			(size_t lvl,	size_t bline);
-std::string embedd			(size_t lvl,	size_t bline,		size_t node);
-std::string embedd_array	(size_t lvl,											size_t arr_size_d1);
-std::string embedd_array	(size_t lvl,	size_t bline,							size_t arr_size_d1);
-std::string embedd_array	(size_t lvl,	size_t bline,		size_t node,		size_t arr_size_d1);
-std::string embedd_dma_array(size_t lvl,	size_t bline,		size_t node,		size_t arr_size_d1);
-std::string embedd_array_2D	(size_t lvl,											size_t arr_size_d1,	size_t arr_size_d2);
-std::string embedd_array_2D	(size_t lvl,	size_t bline,							size_t arr_size_d1,	size_t arr_size_d2);
+std::string embedd								  (
+														size_t lvl,
+														size_t bline);
+std::string embedd								  (
+														size_t lvl,
+														size_t bline,
+														size_t node);
+std::string embedd_array							(
+														size_t lvl,
+														size_t arr_size_d1);
+std::string embedd_array							(
+														size_t lvl,
+														size_t bline,
+														size_t arr_size_d1);
+std::string embedd_array							(
+														size_t lvl,
+														size_t bline,
+														size_t node,
+														size_t arr_size_d1);
+std::string embedd_dma_array						(
+														size_t lvl,
+														size_t bline,
+														size_t node,
+														size_t arr_size_d1);
+std::string embedd_array_2D						 (
+														size_t lvl,
+														size_t arr_size_d1,
+														size_t arr_size_d2);
+std::string embedd_array_2D						 (
+														size_t lvl,
+														size_t bline,
+														size_t arr_size_d1,
+														size_t arr_size_d2);
